# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Bravo API
  version: 1.0.0
  description: OpenAPI documentation for the Bravo backend.
paths:
...
  /api/jobs/:
    get:
      operationId: jobs_list
      description: Lists the jobs where the authenticated user participates as a client
        or organization.
      summary: List jobs
      tags:
      - jobs
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
          description: ''
    post:
      operationId: jobs_create
      description: Creates a new job associated with the authenticated user.
      summary: Create job
      tags:
      - jobs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Job'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Job'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Job'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
          description: ''
  /api/jobs/{uuid}/:
    get:
      operationId: jobs_retrieve
      description: Returns the details of a job accessible to the authenticated user.
      summary: Get job
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - jobs
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
          description: ''
    put:
      operationId: jobs_update
      description: Fully replaces a job accessible to the authenticated user.
      summary: Replace job
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - jobs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Job'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Job'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Job'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
          description: ''
    patch:
      operationId: jobs_partial_update
      description: Partially updates a job accessible to the authenticated user.
      summary: Update job
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - jobs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedJob'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedJob'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedJob'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
          description: ''
    delete:
      operationId: jobs_destroy
      description: Deletes a job accessible to the authenticated user.
      summary: Delete job
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - jobs
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/jobs/{uuid}/messages/:
    get:
      operationId: jobs_messages_retrieve
      description: Gets job chat messages or creates a new message in that chat.
      summary: Manage job messages
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - jobs
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
          description: ''
    post:
      operationId: jobs_messages_create
      description: Gets job chat messages or creates a new message in that chat.
      summary: Manage job messages
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - jobs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Job'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Job'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Job'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
          description: ''
...
  /api/management/jobs/:
    get:
      operationId: management_jobs_list
      tags:
      - management
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManagementJob'
          description: ''
    post:
      operationId: management_jobs_create
      tags:
      - management
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManagementJob'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ManagementJob'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ManagementJob'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementJob'
          description: ''
  /api/management/jobs/{uuid}/:
    get:
      operationId: management_jobs_retrieve
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - management
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementJob'
          description: ''
    put:
      operationId: management_jobs_update
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - management
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManagementJob'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ManagementJob'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ManagementJob'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementJob'
          description: ''
    patch:
      operationId: management_jobs_partial_update
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - management
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedManagementJob'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedManagementJob'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedManagementJob'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementJob'
          description: ''
...
components:
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        uuid:
          type: string
          format: uuid
          readOnly: true
        user:
          type: integer
          readOnly: true
        organization:
          type: integer
        plan_price:
          type: integer
        status:
          $ref: '#/components/schemas/JobStatusEnum'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - organization
      - plan_price
      - updated_at
      - user
      - uuid
    JobStatusEnum:
      enum:
      - pending
      - active
      - inactive
      - suspended
      - completed
      - rejected
      type: string
      description: |-
        * `pending` - Pending
        * `active` - Active
        * `inactive` - Inactive
        * `suspended` - Suspended
        * `completed` - Completed
        * `rejected` - Rejected
    ManagementJob:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        uuid:
          type: string
          format: uuid
          readOnly: true
        user:
          type: string
          format: uuid
        organization:
          type: string
          format: uuid
        plan_price:
          type: integer
        status:
          $ref: '#/components/schemas/JobStatusEnum'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - organization
      - plan_price
      - updated_at
      - user
      - uuid
...
    PatchedJob:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        uuid:
          type: string
          format: uuid
          readOnly: true
        user:
          type: integer
          readOnly: true
        organization:
          type: integer
        plan_price:
          type: integer
        status:
          $ref: '#/components/schemas/JobStatusEnum'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    PatchedManagementJob:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        uuid:
          type: string
          format: uuid
          readOnly: true
        user:
          type: string
          format: uuid
        organization:
          type: string
          format: uuid
        plan_price:
          type: integer
        status:
          $ref: '#/components/schemas/JobStatusEnum'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT